requirements.txt
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
PyYAML==6.0.1
typeguard==2.13.3
pydantic==2.4.2
aiohttp==3.9.1
python-multipart==0.0.6
httpx==0.25.2
facebook-sdk==3.1.0
instagram-private-api==1.6.0
pillow==10.0.0
